[Priest] More refactoring, removal of lost Artifact stuff

+++ engine/class_modules/priest/sc_priest.cpp
@@ -442,7 +442,8 @@ void mangazas_madness( special_effect_t& effect )
 
   if ( priest->active_items.mangazas_madness )
   {
-    priest->cooldowns.mind_blast->charges += priest->active_items.mangazas_madness->driver()->effectN( 1 ).base_value();
+    priest->cooldowns.mind_blast->charges += 
+      (int)(float) priest->active_items.mangazas_madness->driver()->effectN( 1 ).base_value();
   }
 }
